Implement Fragmented WebVTT vtte/vttc
As #1638 reported, we do not support some (or all?) ways of
communicating webvtt subtitles in mp4 files.

I initially assumed that webvtt embedded in an mp4 file worked like for
TTML (a format we're more used to): the plain subtitle format directly
inserted in an `mdat` box.

Turns out, some contents (like https://demo.unified-streaming.com/k8s/features/stable/video/tears-of-steel/tears-of-steel-wvtt.ism/.mpd),
actually rely on the metadata from other ISOBMFF boxes (mainly the
tfdt and trun boxes) to provide timing information with the `mdat`
reserved for text, identifiers and style information in a new binary
format following the ISOBMFF way of doing things (new webvtt-specific
boxes).

Weirdly enough, that format does not look at all like the WEBVTT we're
used to beside the fact that it uses the same identifier and "settings"
concept.

---

As now our subtitle parser has to have the context of the whole mp4 file
(and not of the mp4 segment), and as that parser also has to rely on
state (a `timescale` value) provided by an initialization segment, I had
to update the architecture of how subtitles are communicated: they can
now be communicated as string or `BufferSource` types (the latter
leading the text encoding detection), and a supplementary `timescale`
argument (defaulting to `1`) is now always provided to parsers.

The vast majority of parsers now do not make use of that `timescale`
value which is kind of ugly though, we may want to find a better
solution.