Fix: for Safari directfile, `play()` even if seek is not finished
We noticed an issue on some devices, for the moment only seen with
Safari desktop directfile (e.g. playing an HLS content in directfile
through the RxPlayer) where a deadlock is going on after we perform
the initial seek

- Safari waits for playback to begin (e.g. by calling the `play()`
  method on the video tag) before loading content and resetting the
  video element's `seeking` attribute to `false`

- The RxPlayer waits for the `seeking` attribute to be set to `false`
  before calling `play()`, to ensure it has enough data to play
  reliably.

As this is an optimization and as in "directfile", the browser is the
one often in charge to know when it has enough data to play anyway, I
propose here to just disable this behavior for Safari desktop when
playing a directfile content.